Arrays are not pointers

+++ 0x05-pointers_arrays_strings/8-print_array.c
@@ -12,7 +12,7 @@
 
 void print_array(int *a, int n)
 {
-	unsigned long int i = 0;
+	int i = 0;
 
 	while (i < n - 1)
 	{
